完成配置文件的填充

+++ src/main/java/com/tac/guns/event/ammo/DestroyGlassBlock.java
@@ -1,6 +1,7 @@
 package com.tac.guns.event.ammo;
 
 import com.tac.guns.api.event.AmmoHitBlockEvent;
+import com.tac.guns.config.common.AmmoConfig;
 import com.tac.guns.entity.EntityBullet;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
@@ -17,8 +18,7 @@ public class DestroyGlassBlock {
         BlockState state = event.getState();
         BlockPos pos = event.getHitResult().getBlockPos();
         EntityBullet ammo = event.getAmmo();
-        if (state.getMaterial() == Material.GLASS) {
-            // TODO 打碎玻璃（可以给个 Config 开关）
+        if (AmmoConfig.DESTROY_GLASS.get() && state.getMaterial() == Material.GLASS) {
             level.destroyBlock(pos, false, ammo.getOwner());
         }
     }
